Move backend fetch wrappers into lib/fetch.
